+++ js/utils/get-random-number.js
@@ -1,0 +1,10 @@
+// source https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+function getRandomNumber(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+getRandomNumber(10, 20);
+
+export { getRandomNumber };
